in_callback flag is problematic.
I don't like the in_callback flag as it requires clients with it
set to be purged upon specific locations. Also it would require
reproducing a lot of setting of that flag around the API, in
future functions other than sukat_sock. Instead I'll add the
deletion list to sukat_sock. With it we need only a few points in
the code (in end of read and in ctx_destroy) where we need to
clear the deletion list.