Added start of a BGP-library.
Noticed that the same pattern of how to handle disconnects and
destroys when in callback begins to appear again. I'll cut BGP a
bit short for now and work on making an internal helper API for
operating inside callbacks. The idea is to register close, free
and checking/setting of in_callback flags to it. After "callback
sections", the helper can be used to free any dangling non-free'd
data.