Changed API to separate endpoints and contexts.
Made the sukat_sock_endpoint_add call somewhat mandatory as the
socket is no longer created when the socket context is created.
Instead each server and client must explicitly add its
server/client fds by sukat_sock_endpoint_add. This returns a
handle to that endpoint, which eliminates the need to follow up
on the endpoints in any way in the socket context itself.

Also the socket types etc. were migrated to the endpoint context.
BGP tests disabled for a while, as the whole bgp library needs
some redesign.

Might still want to separate the socket parameters from the
endpoint parameters as they are totally different.