Accepted, added and disconnected peers.
Refactored the conn_cb boolean for the event type to an enum.
This helps with the user of the library to distinguish if he
should expect the caller_ctx to be his main context or an already
allocated peer-specific context.

Fixed the BGP library to use the new endpoint-oriented
sukat_sock library. Now the library will always act as server,
adding peers. This is more inline with the official RFC anyway.