Keepalive and peer stringify
Added support for the keepalive-message to the BGP library. Also
RFC mandated that open-messages are replied to by a keepalive
message so added sending of them after an open-message. Also
theres a keepalive cb so peer timing can be tracked.

Added debugging macros that take the peer as an argument. This
makes logs better as the peer is automatically formatted to a
string for easy tracking.