BGP UPDATE message.
Added support for the UPDATE message in the BGP protocol. The
withdrawn and reachable routes aren't converted for the user by
the library but all other data is formatted to a easier to handle
format of a linked list, where all values have been converted to
host byte order.

There might be some problems arising from using variadic array
members in unions. I'll remove those from the unions and use
separate pointers for such structures if need be.