Sukat BGP against Quagga BGP.
Implemented a lot of parsing fixes discovered against the quagga
BGP server. For example the length use for path attributes was
broken as the extended flag wasn't taken into consideration and
the length was presumed to not even be present.

Implemented the update_cb from bgp_query so it logs out all
data given in the update.