Sending and receiving for message boundaries
Refactored and implemented the sukat_sock parts that use messages
using message boundaries (DGRAM and SEQPACKET). Refactoring the
connectionless sending and receiving yielded most of the parts to
be re-usable by the SEQPACKET parts.

Testing for SEQPACKET is coming in the next socket. DGRAM-side
would be nice to have some more unit tests too.