Logging improvements.
Refactore the logging of peers and sockaddr-data to more shared
functions, so duplicacy could be removed.

For some reason on my server the tests with multiple message fail
on AF_UNIX SOCK_DGRAM tests. I reduced their size to make the
socket not block on them. It might be nice do to them as best effort,
where the messages are sent until it blocks.