Google benchmark with AF_INET and AF_UNIX.
Added tests for SOCK_STREAM with AF_INET and AF_UNIx with
different message lengths. I'm not that good with C++ so the
current arrangement seems a bit off. The range_y() parameter
should be the domain, but I'm hesitant to drop the fixtures as I
don't want socket creation overhead to affect the tests.

There is still a mysterious bug that causes netdoges splicing to
not work at all on one of my machines. Also sometimes on the
other machine I have to start a new urxvt terminal and then it
suddendly works again, if it failed previously with invalid
argument.